corrections

+++ 0x10-variadic_functions/3-print_all.c
@@ -14,6 +14,8 @@ void print_all(const char * const format, ...)
 
 	va_list list;
 
+
+
 	va_start(list, format);
 
 	if (format)
@@ -29,7 +31,7 @@ void print_all(const char * const format, ...)
 				printf("%s%d", sep, va_arg(list, int));
 				break;
 			case 'f':
-				printf("%s%d", sep, va_arg(list, int));
+				printf("%s%f", sep, va_arg(list, double));
 				break;
 			case 's':
 				str = va_arg(list, char *);
